requirements.txt
streamlit
openai==1.96.1
ipython
python-dotenv 